Fix issue with Left's implicit conversion to eithers.
The actual value of the Left was not passed on to the either
constructor, leading to subtle bugs when relying on the implicit
conversion operator e.g. in match expressions.

In addition, corrected a noexcept condition in sum_type, where a
constructor parameter was not propagated in the noexcept specification
(though it was in the actual constructor call).