Attempt to fix travis build error for clang.
For whatever reason, the buildit script for libcxx fails with:

  .\buildit: 117: Bad substitution

Since the revision of libcxx being checked out is exactly  the same,
the assumption is that something must have changed in the travis
environment. Some research shows that a common cause of substitution
errors is running a shell script in an incompatible shell, hence the
explicit bash invocation here.